Add extra checks if user has permission to edit/delete comment

+++ app/Http/Livewire/CommentItem.php
@@ -28,7 +28,17 @@ class CommentItem extends Component
 
     public function deleteComment()
     {
+        $user = auth()->user();
+        if (!$user) {
+            return $this->redirect('/login');
+        }
+
+        if ($this->comment->user_id != $user->id) {
+            return response('You are not allowed to perform this action', 403);
+        }
+
         $id = $this->comment->id;
+
         $this->comment->delete();
         $this->emitUp('commentDeleted', $id);
     }
